test file reorganization corrected and documented
