+++ tests/cli/conftest.py
@@ -1,0 +1,22 @@
+"""Pytest configuration for CLI tests."""
+
+import pytest
+from .helpers import get_installed_apps
+
+
+def pytest_generate_tests(metafunc):
+    """Dynamically parametrize vba_app based on command line options."""
+    if "vba_app" in metafunc.fixturenames:
+        # Get selected apps from command line
+        apps_option = metafunc.config.getoption("--apps")
+        if apps_option.lower() == "all":
+            selected_apps = ["excel", "word", "access"]
+        else:
+            selected_apps = [app.strip().lower() for app in apps_option.split(",")]
+            valid_apps = ["excel", "word", "access"]
+            invalid_apps = [app for app in selected_apps if app not in valid_apps]
+            if invalid_apps:
+                raise ValueError(f"Invalid apps: {invalid_apps}. Valid options: {valid_apps}")
+
+        apps = get_installed_apps(selected_apps=selected_apps)
+        metafunc.parametrize("vba_app", apps, ids=lambda x: f"{x}-vba")
